Migrate requestlimit to caddy2
