+++ go/caddy/caddy/caddy/caddymain/run.go
@@ -45,7 +45,6 @@ import (
 	pflog "github.com/inverse-inc/go-utils/log"
 	_ "github.com/inverse-inc/packetfence/go/caddy/forwardproxy"
 	_ "github.com/inverse-inc/packetfence/go/caddy/pfpki"
-	_ "github.com/inverse-inc/packetfence/go/caddy/requestlimit"
 )
 
 func init() {
